optimize unit test.
