+++ hadoop-hdfs-project/hadoop-hdfs-rbf/src/test/java/org/apache/hadoop/hdfs/server/federation/router/TestRouterRpc.java
@@ -159,7 +159,6 @@ import org.apache.hadoop.thirdparty.com.google.common.collect.Maps;
  * Tests covering the functionality of RouterRPCServer with
  * multi nameServices.
  */
-@SuppressWarnings("checkstyle:VisibilityModifier")
 public class TestRouterRpc {
 
   private static final Logger LOG =
@@ -193,7 +192,7 @@ public class TestRouterRpc {
   private NamenodeContext namenode;
 
   /** Client interface to the Router. */
-  protected ClientProtocol routerProtocol;
+  private ClientProtocol routerProtocol;
   /** Client interface to the Namenode. */
   private ClientProtocol nnProtocol;
 
@@ -204,7 +203,7 @@ public class TestRouterRpc {
   private NamenodeProtocol nnNamenodeProtocol1;
 
   /** Filesystem interface to the Router. */
-  protected FileSystem routerFS;
+  private FileSystem routerFS;
   /** Filesystem interface to the Namenode. */
   private FileSystem nnFS;
 
@@ -2386,5 +2385,46 @@ public class TestRouterRpc {
       fileSystem1.delete(new Path(testPath1), true);
       fileSystem1.delete(new Path(testPath2), true);
     }
+  }
+
+  @Test
+  public void testCallerContextNotResetByAsyncHandler() throws IOException {
+    GenericTestUtils.LogCapturer auditLog =
+        GenericTestUtils.LogCapturer.captureLogs(FSNamesystem.AUDIT_LOG);
+    String dirPath = "/test";
+
+    // The reason we start this child thread is that CallContext use InheritableThreadLocal.
+    Thread t1 = new Thread(() -> {
+      // Set flag async:true.
+      CallerContext.setCurrent(
+          new CallerContext.Builder("async:true").build());
+      // Issue some RPCs via the router to populate the CallerContext of async handler thread.
+      for (int i = 0; i < 10; i++) {
+        try {
+          routerProtocol.mkdirs(dirPath, new FsPermission("755"), false);
+          assertTrue(verifyFileExists(routerFS, dirPath));
+          routerProtocol.delete(dirPath, true);
+          assertFalse(verifyFileExists(routerFS, dirPath));
+        } catch (Exception e) {
+          throw new RuntimeException(e);
+        }
+      }
+
+      // The audit log should contains async:true.
+      assertTrue(auditLog.getOutput().contains("async:true"));
+      auditLog.clearOutput();
+      assertFalse(auditLog.getOutput().contains("async:true"));
+    });
+
+    t1.start();
+    try {
+      t1.join();
+    } catch (InterruptedException e) {
+      throw new RuntimeException(e);
+    }
+
+    routerProtocol.getFileInfo(dirPath);
+    // The audit log should not contain async:true.
+    assertFalse(auditLog.getOutput().contains("async:true"));
   }
 }
